Reorganise parser classes into new util package
