+++ src/main/java/duke/command/Command.java
@@ -1,7 +1,7 @@
 package duke.command;
 
 import duke.DukeException;
-import duke.Parser;
+import duke.util.Parser;
 import duke.Storage;
 import duke.Ui;
 import duke.task.TaskList;
