+++ src/main/java/duke/task/Deadline.java
@@ -1,9 +1,13 @@
 package duke.task;
 
-import duke.DateParser;
+import duke.util.DateParser;
 import duke.DukeException;
 
 public class Deadline extends Task {
+    /**
+     * The deadline this task must be completed by.
+     * It is a String to allow the user to custom dates such as 'the day after tomorrow'
+     */
     protected String date;
 
     public Deadline(String desc, String date) {
