+++ src/main/java/duke/util/Parser.java
@@ -1,4 +1,6 @@
-package duke;
+package duke.util;
+
+import duke.DukeException;
 
 import java.util.Arrays;
 import java.util.HashMap;
